Introduce improved mechanism for util.Stopper.
In new conception, there is a draining phase. Any component which does
any unit of work as a "first mover" increments a drain count via
Stopper.Inc() and then decrements via Stopper.Dec() on completion. The
units of work which count are things like received external requests
and periodic work done in the gossip loop. While any of these things are
active, the stopper is prevented from initiating stoppage. However, all
first movers get false when calling Inc(), meaning they don't allow new
work to be scheduled.

The system quiesces when all draining is done at which point the stopper
signals all components to stop via ShouldStop() and they proceed to stop.

The idea behind the change is we thread an entire running system together
with a single stopper and are able to stop a complex system with more
confidence by allowing all activity to drain and then shutting down all
components.

The idea seems to have good results in terms of simplifying some of the
ugly stopping code which has been introduced of late. The gossip code got
a lot simpler as well.

Added a quit handler in the _admin REST API and corresponding CLI
command to quit a node. The draining we have isn't fully featured here,
but this is a start.